requirements.txt
flask
flask-cors
flask-pymongo
pyjwt
requests
python-dotenv
dnspython
